refactor(index.js): testing needed to not export from where the app starts
