+++ src/components/map/index.js
@@ -1,5 +1,6 @@
 export { default as Node } from './node.jsx'
 export { default as NodeObject } from './nodeObject.jsx'
 export { default as NodeTmpl } from './node.tmpl.jsx'
-export { default as NodeObjectTmpl } from './nodeObject.tmpl.jsx'
 export { default as MapTmpl } from './map.tmpl.jsx'
+export { default as NodeObjectTmpl } from './nodeObject.tmpl.jsx'
+export { default as Toolbar } from './Toolbar.jsx'
