+++ src/components/map/map.tmpl.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { Node } from './';
+import { Node, Toolbar } from './';
 import SideBar from '../sideBar/sideBar.jsx';
-import Toolbar from './Toolbar';
 //makeDraggable should be configured in the parent and pass teh whole thing in and then wouldn't need the whole this in here
 
 function MapTmpl(props) {
